Move WB tests to the curve_maps
